py-server/requirements.txt
# Core Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Audio Processing
pydub>=0.25.1
librosa>=0.10.1
soundfile>=0.12.1
noisereduce>=3.0.0

# AI & ML - PyTorch (CPU version, use cuda121 for GPU)
# Install with: pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchaudio>=2.0.0

# WhisperX - Unified transcription + alignment + diarization
# Install with: pip install git+https://github.com/m-bain/whisperx.git
whisperx @ git+https://github.com/m-bain/whisperx.git

# Language Detection
langdetect>=1.0.9

# Utilities
numpy>=1.24.0
scipy>=1.11.0
requests>=2.31.0
